NodeManager/requirements.txt
azure-common==1.1.28
azure-core==1.26.3
azure-identity==1.12.0
azure-mgmt-compute==29.1.0
azure-mgmt-core==1.3.2
azure-storage-blob==12.15.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.2
numpy==1.24.2
oauthlib==3.2.2
pandas==1.5.3
pipreqs==0.4.11
portalocker==2.7.0
psutil==5.9.4
psycopg2-binary==2.9.5
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==2.0.9

